Debugged so it works now. Had issues because packages weren't declared correctly

+++ app/src/main/java/com/example/financehub/data/database/Expense.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class Expense(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val title: String,
-    val amount: Double
+    val amount: Int
 )
